feat: added protected routes, en el back y en el front

+++ frontend/src/components/WorkoutDetails.jsx
@@ -6,20 +6,27 @@ import { es } from 'date-fns/locale';
 
 // Custom hooks
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // Actions creators
 import { deleteWorkout } from '../context/actions/workouts.actions';
 
 const WorkoutDetails = ({ id, title, reps, load, createdAt }) => {
 
+  const {state:{user}} = useAuthContext();
   const { dispatch } = useWorkoutsContext();
 
   async function handleDelete(id) {
+
+    if(!user) {
+      return;
+    }
 
     const res = await fetch(`http://localhost:5400/api/workouts/${id}`, {
       method: 'DELETE',
       headers: {
         "Content-Type": "application/json",
+        authorization: `Bearer ${user.token}`,
       }
     });
